REFACTOR: Renamed components in `./app` directory to use kebab case and renamed folders to `_components` as per conventions
