+++ app/albums/[id]/_components/album-detail-content.tsx
