--- app/albums/[id]/components/album-detail-content.tsx
@@ -1,255 +1,0 @@
-"use client";
-
-import { useState } from "react";
-import Image from "next/image";
-import { useRouter } from "next/navigation";
-import { Pencil, Trash2, MoreHorizontal } from "lucide-react";
-import { toast } from "sonner";
-import type { AlbumDetail } from "../../../../types/music/album-detail";
-import useLoadImage from "@/hooks/useLoadImage";
-import { formatArtists } from "@/lib/music/format-artists";
-import { toSongsWithAlbum } from "@/lib/mappers/song";
-import SongsGrid from "@/components/song/songs-grid";
-import { useUser } from "@/hooks/useUser";
-import renameAlbum from "@/actions/album/rename-album";
-import deleteAlbum from "@/actions/album/delete-album";
-import { RenameAlbumSchema } from "@/schemas/albums/rename-album.schema";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { ROUTES } from "@/routes";
-
-interface AlbumDetailContentProps {
-  album: AlbumDetail;
-}
-
-/**
- * Detailed content for an individual album.
- * Displays album information, cover art, and a grid of songs.
- * Provides administrative features (rename, delete) if the user is the owner.
- *
- * @param album - The detailed album object to display.
- * @author Maruf Bepary
- */
-const AlbumDetailContent: React.FC<AlbumDetailContentProps> = ({ album }) => {
-  const router = useRouter();
-  const { user } = useUser();
-
-  const [isRenameDialogOpen, setIsRenameDialogOpen] = useState(false);
-  const [newTitle, setNewTitle] = useState(album.title);
-  const [isRenaming, setIsRenaming] = useState(false);
-  const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false);
-  const [isDeleting, setIsDeleting] = useState(false);
-
-  const isOwner = user?.id === album.uploaderId;
-
-  const songsWithAlbum = toSongsWithAlbum(album);
-  const imageUrl = useLoadImage(album.coverImagePath);
-  const artistNames = formatArtists(album);
-  const releaseYear = album.releaseDate
-    ? new Date(album.releaseDate).getFullYear()
-    : null;
-
-  /**
-   * Calls the `deleteAlbum` server action, redirects to /albums on success,
-   * and shows a toast notification for both success and failure cases.
-   *
-   * @author Maruf Bepary
-   */
-  const onDelete = async () => {
-    setIsDeleting(true);
-    try {
-      const { ok, error } = await deleteAlbum(album.id);
-      if (ok) {
-        toast.success("Album deleted");
-        router.push(ROUTES.ALBUMS.path);
-        router.refresh();
-      } else {
-        toast.error(error || "Failed to delete album");
-      }
-    } catch {
-      toast.error("An error occurred while deleting the album");
-    } finally {
-      setIsDeleting(false);
-      setIsDeleteDialogOpen(false);
-    }
-  };
-
-  /**
-   * Validates the new title with RenameAlbumSchema, then calls the `renameAlbum`
-   * server action. Dismisses the dialog and refreshes the route on success.
-   * No-ops if the title is unchanged or empty.
-   *
-   * @author Maruf Bepary
-   */
-  const onRename = async () => {
-    const trimmed = newTitle.trim();
-    if (!trimmed || trimmed === album.title) {
-      setIsRenameDialogOpen(false);
-      return;
-    }
-    const parsed = RenameAlbumSchema.safeParse({ albumId: album.id, newTitle: trimmed });
-    if (!parsed.success) {
-      toast.error(parsed.error.issues[0]?.message ?? "Invalid album title");
-      return;
-    }
-    setIsRenaming(true);
-    try {
-      const success = await renameAlbum(album.id, trimmed);
-      if (success) {
-        toast.success("Album renamed");
-        router.refresh();
-        setIsRenameDialogOpen(false);
-      } else {
-        toast.error("Failed to rename album");
-      }
-    } catch {
-      toast.error("Failed to rename album");
-    } finally {
-      setIsRenaming(false);
-    }
-  };
-
-  return (
-    <div className="flex flex-col gap-y-6 px-6 pb-6">
-      {/* Album Header */}
-      <div className="flex flex-col items-center gap-6 sm:flex-row sm:items-end">
-        <div className="relative h-48 w-48 shrink-0 overflow-hidden rounded-md shadow-lg sm:h-56 sm:w-56">
-          <Image
-            src={imageUrl ?? "/images/music-placeholder.png"}
-            fill
-            sizes="(max-width: 640px) 192px, 224px"
-            alt={album.title}
-            className="object-cover"
-            priority
-          />
-        </div>
-        <div className="flex flex-col items-center gap-y-2 sm:items-start">
-          <h1 className="text-3xl font-bold sm:text-4xl">{album.title}</h1>
-          {isOwner && (
-            <DropdownMenu>
-              <DropdownMenuTrigger asChild>
-                <Button variant="outline" size="sm">
-                  <MoreHorizontal className="size-4 mr-2" />
-                  Options
-                </Button>
-              </DropdownMenuTrigger>
-              <DropdownMenuContent align="start">
-                <DropdownMenuItem
-                  onClick={() => {
-                    setNewTitle(album.title);
-                    setIsRenameDialogOpen(true);
-                  }}
-                  disabled={isRenaming}
-                >
-                  <Pencil className="mr-2 size-4" />
-                  Rename
-                </DropdownMenuItem>
-                <DropdownMenuItem
-                  className="text-destructive focus:text-destructive"
-                  onClick={() => setIsDeleteDialogOpen(true)}
-                  disabled={isDeleting}
-                >
-                  <Trash2 className="mr-2 size-4" />
-                  Delete Album
-                </DropdownMenuItem>
-              </DropdownMenuContent>
-            </DropdownMenu>
-          )}
-          <p className="text-sm text-muted-foreground">
-            {artistNames}
-            {releaseYear && ` • ${releaseYear}`}
-            {` • ${album.songs.length} ${album.songs.length === 1 ? "track" : "tracks"}`}
-          </p>
-        </div>
-      </div>
-
-      {/* Track Listing */}
-      <div className="flex flex-col gap-y-2">
-        <h2 className="text-xl font-semibold">Tracks</h2>
-        <SongsGrid songs={songsWithAlbum} />
-      </div>
-
-      {/* Delete Confirmation Dialog */}
-      <Dialog open={isDeleteDialogOpen} onOpenChange={setIsDeleteDialogOpen}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Delete Album</DialogTitle>
-            <DialogDescription>
-              Are you sure you want to delete &quot;{album.title}&quot;? All tracks in this album will also be permanently deleted. This action cannot be undone.
-            </DialogDescription>
-          </DialogHeader>
-          <DialogFooter>
-            <Button
-              variant="outline"
-              onClick={() => setIsDeleteDialogOpen(false)}
-              disabled={isDeleting}
-            >
-              Cancel
-            </Button>
-            <Button
-              variant="destructive"
-              onClick={onDelete}
-              disabled={isDeleting}
-            >
-              {isDeleting ? "Deleting..." : "Delete"}
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
-
-      {/* Rename Dialog */}
-      <Dialog
-        open={isRenameDialogOpen}
-        onOpenChange={(open) => {
-          if (!open) setNewTitle(album.title);
-          setIsRenameDialogOpen(open);
-        }}
-      >
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Rename Album</DialogTitle>
-            <DialogDescription>
-              Enter a new name for this album.
-            </DialogDescription>
-          </DialogHeader>
-          <Input
-            value={newTitle}
-            onChange={(e) => setNewTitle(e.target.value)}
-            placeholder="Album title"
-          />
-          <DialogFooter>
-            <Button
-              variant="outline"
-              onClick={() => setIsRenameDialogOpen(false)}
-              disabled={isRenaming}
-            >
-              Cancel
-            </Button>
-            <Button
-              onClick={onRename}
-              disabled={isRenaming || !newTitle.trim()}
-            >
-              Rename
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
-    </div>
-  );
-};
-
-export default AlbumDetailContent;
