+++ app/artists/[id]/_components/artist-detail-content.tsx
