+++ app/artists/_components/artists-content.tsx
