+++ app/artists/_components/artists-header.tsx
