+++ app/favourites/page.tsx
@@ -1,6 +1,6 @@
 import getFavouriteSongs from "@/actions/playlist/get-favourite-songs";
 import Header from "@/components/header";
-import FavouritesContent from "./components/favourites-content";
+import FavouritesContent from "./_components/favourites-content";
 
 export const revalidate = 0;
