+++ app/playlists/[id]/_components/playlist-detail-content.tsx
